backend/app/services/ml_service/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
numpy==2.1.3
pillow==11.0.0
tensorflow==2.16.1
python-multipart==0.0.9
requests==2.32.3

